Move resource-reducer into its own file
